Replaced Square class to function and Added handleClick
It is better to write functions for components that only contain render method and don't have their own state.

AND

In handleClick method, we are creating and changing data without mutation (using Slice method and then changing value to X). (This is because it helps to re-render because immutable items are easy to track and also this simplifies complex features)